Fixed some minor bugs with code blocks
Fixed early termination of indented code blocks -- they were looking
for empty strings (which can occur inside a code block if the leading
indent is there), it now checks for 4 spaces instead. Also got rid of
the extra leading whitespace emitted for these code blocks (like
multimarkdown does). Moved the newline to the end instead of the
beginning of a fenced code block -- it was always emitting an empty
leading line on those blocks.