Another code block corner case fixed
When a fenced code block has code on the same line as the initial 3
back quotes, a newline was not getting emitted. I don't think this is
normal, but it is now fixed.
Also updated the tests to expect the output consistent with
multimarkdown, not what it happened to emit before.